server/pyproject.toml
[tool.poetry]
name = "all-the-reviews-server"
version = "0.1.0"
description = ""
authors = ["misung Kim <mizzkkim7@gmail.com>"]
readme = "README.md"
packages = [{include = "all_the_reviews_server"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
pillow = "^9.4.0"
django-allauth = "^0.52.0"
dj-rest-auth = "^3.0.0"
google-api-python-client = "^2.79.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
